+++ aks_cluster/provider.tf
@@ -1,7 +1,12 @@
 terraform {
   required_version = ">= 0.12"
-  required_providers {
-    azurerm = ">=1.43.0"
-    azuread = ">= 0.3"
-  }
+}
+
+provider "azurerm" {
+  version = "~> 2.0.0"
+  features {}
+}
+
+provider "azuread" {
+  version = "~> 0.4"
 }