+++ aks_node_pool/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## aks_node_pool-v0.2.0
+
+### Fixed
+
+- added lifecycle ignore for node_taints
+
+### Updated
+
+- Upgraded azurerm provider version to 2.0.0
+
+___
+
 ## aks_node_pool-v0.1.0
 
 - Initial release for the aks_node_pool module